+++ source/extensions/isaacsim.ros2.nodes/python/tests/test_camera_info.py
@@ -47,9 +47,7 @@ SAVE_DEPTH_IMAGES_AS_TEST = False
 SAVE_DEPTH_IMAGES_AS_GOLDEN = False
 
 
-# Having a test class derived from omni.kit.test.AsyncTestCase declared on the root of module will make it auto-discoverable by omni.kit.test
 class TestRos2CameraInfo(ROS2TestCase):
-    # Before running each test
     async def setUp(self):
         await super().setUp()
 
@@ -66,7 +64,6 @@ class TestRos2CameraInfo(ROS2TestCase):
 
         pass
 
-    # After running each test
     async def tearDown(self):
 
         self._timeline.stop()